// src/types/components.ts
export type buttonTypes = "basic" | "submit" | "reset" | "primary1" | "primary2" | "section1" | "section2" | "bd1" | "bd2" | null;

export type SuccessStatProps = {
  cardNumber: 1 | 2 | 3 | 4 | 5;
};

export type blogArrayTypes = {
  img: string;
  category: string;
  header: string;
  text: string;
  author: string;
  date: string;
  className: string;
};

export type testimonialsArrayTypes = {
  img: string;
  name: string;
  text: string;
  stars: string;
  reviews: string;
};